Day 1 of PyQT5 Practice
Learned a lot today, about the boilerplate code needed for a basic application, how widgets work a bit. Learned about how to make the widgets interactive and how they send signals and each can be connected to a slot(a function you define or a preexisting function) or even multiple different slot. And how widgets can connect directly to each other as well!
Overall a very productive day and I'm very happy with the progress!